+++ src/audio/CMakeLists.txt
@@ -183,6 +183,7 @@ if(CONFIG_IPC_MAJOR_3)
 	set(tdfb_sources tdfb/tdfb_ipc3.c)
 	set(tdfb_sources multiband_drc/multiband_drc_ipc3.c)
 	set(dcblock_sources dcblock/dcblock_ipc3.c)
+	set(mux_sources mux/mux_ipc3.c)
 elseif(CONFIG_IPC_MAJOR_4)
 	set(volume_sources volume/volume.c volume/volume_generic.c volume/volume_ipc4.c)
 	set(src_sources src/src.c src/src_ipc4.c src/src_generic.c)
@@ -191,6 +192,7 @@ elseif(CONFIG_IPC_MAJOR_4)
 	set(tdfb_sources tdfb/tdfb_ipc4.c)
 	set(tdfb_sources multiband_drc/multiband_drc_ipc4.c)
 	set(dcblock_sources dcblock/dcblock_ipc4.c)
+	set(mux_sources mux/mux_ipc4.c)
 endif()
 set(mixer_sources ${mixer_src})
 set(asrc_sources asrc/asrc.c asrc/asrc_farrow.c asrc/asrc_farrow_generic.c)
