+++ test/cmocka/src/audio/mux/CMakeLists.txt
@@ -10,6 +10,7 @@ add_library(
 	audio_mux
 	STATIC
 	${PROJECT_SOURCE_DIR}/src/audio/mux/mux.c
+	${PROJECT_SOURCE_DIR}/src/audio/mux/mux_ipc3.c
 	${PROJECT_SOURCE_DIR}/src/audio/mux/mux_generic.c
 	${PROJECT_SOURCE_DIR}/src/audio/component.c
 	${PROJECT_SOURCE_DIR}/src/audio/data_blob.c
@@ -48,3 +49,8 @@ cmocka_test(
 	demux_copy
 	demux_copy.c
 )
+
+target_include_directories(audio_mux PRIVATE ${PROJECT_SOURCE_DIR}/src/audio)
+target_include_directories(mux_get_processing_function PRIVATE ${PROJECT_SOURCE_DIR}/src/audio)
+target_include_directories(mux_copy PRIVATE ${PROJECT_SOURCE_DIR}/src/audio)
+target_include_directories(demux_copy PRIVATE ${PROJECT_SOURCE_DIR}/src/audio)
